requirements.txt
flask==3.0.3
flask-cors==4.0.1
pydantic==2.7.4
pydantic-settings==2.4.0
werkzeug==3.0.4
openai==1.43.0
spacy==3.7.5
scikit-learn==1.5.1
pdfminer.six==20240706
python-docx==1.1.2
rapidfuzz==3.9.7
numpy==1.26.4
scipy==1.13.1
reportlab==4.2.5
pytest==8.3.2